管理者とオーナーの権限機能を追加

+++ app/Http/Controllers/PostController.php
@@ -89,6 +89,8 @@ class PostController extends Controller
     public function edit(Post $post)
     {
         //
+        // update, destroyでも同様に
+        $this->authorize('edit', $post);
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -102,6 +104,7 @@ class PostController extends Controller
     public function update(Request $request, Post $post)
     {
         //
+        $this->authorize('edit', $post);
         $post->title = $request->title;
         $post->body = $request->body;
         $post->save();
@@ -117,6 +120,7 @@ class PostController extends Controller
     public function destroy(Post $post)
     {
         //
+        $this->authorize('edit', $post);
         $post->delete();
         return redirect('posts');
     }
